ENTRY FUNCTION, CHOICE EXIT
With a little bit of experimentation, I realized I needed to add some handling within the GOTO routine to cover all the bases for the choices framework.

ENTRYFCN is a property placed on a choice room which will attempt to execute a provided function before the room is described. This is useful for if you want to have functions at the start of each story branch to register (within global variables) the choices being made.

CHOICE EXIT is a property which will simply redirect to the room provided. This is useful for if you want to utilize a ENTRY function, but that's basically it.